requirements.txt
requests>=2.31.0
pandas>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.2
gspread>=6.0.0
google-auth>=2.25.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
